// utils/enums/routes.ts
enum Screens {
  RegistrationScreen = "Registration",
  LoginScreen = "Login",
  Home = "Home",
  HomeStack = "HomeStack",
  Posts = "Posts",
  Profile = "Profile",
  CreatePosts = "Create post",
  Comments = "Comments",
  Camera = "Camera",
  Map = "Map",
}

enum ScreensTitles {
  Posts = "Publications",
  Profile = "Profile",
  CreatePosts = "Create post",
  Comments = "Comments",
  Camera = "Camera",
  RegistrationScreen = "Registration",
  LoginScreen = "Login",
  Home = "Home",
}

export { Screens, ScreensTitles };